js-demo: get rid of react-async-component
