+++ crates/wasm/js-demo/js/initialClusters.ts
@@ -1,0 +1,52 @@
+import { Cluster } from '../../pkg';
+export const initialClusters: Cluster[] = [
+    {
+        id: "cluster-1",
+        cites: [
+            { id: "citekey" }
+        ],
+    },
+    {
+        id: "cluster-2",
+        cites: [
+            { id: "citekey2" }
+        ],
+    },
+    {
+        id: "cluster-3",
+        cites: [
+            { id: "citekey", locator: "56" }
+        ],
+    },
+    {
+        id: "cluster-4",
+        cites: [
+            { id: "r1" }
+        ]
+    },
+    {
+        id: "cluster-5",
+        cites: [
+            { id: "ysuf1" }
+        ]
+    },
+    {
+        id: "cluster-6",
+        cites: [
+            { id: "ysuf2" }
+        ]
+    },
+    {
+        id: "cluster-7",
+        cites: [
+            { id: "ysuf1" }
+        ]
+    },
+    {
+        id: "cluster-8",
+        cites: [
+            { id: "r7" }
+        ]
+    },
+];
+
