+++ crates/wasm/js-demo/package.json
@@ -43,7 +43,6 @@
     "graphlib-dot": "^0.6.2",
     "immer": "^3.1.3",
     "react": "^16.8.2",
-    "react-async-component": "^2.0.0",
     "react-dom": "^16.8.2",
     "react-select": "^3.0.4",
     "safe-types": "^4.20.0",
